integration/requirements.txt
requests
selenium
psycopg2-binary
